+++ lib/model/request/content/requests/movie_request.dart
@@ -1,7 +1,7 @@
 import 'package:ombiapp/contracts/media_content_request.dart';
 
-class MovieRequestPodo extends MediaContentRequest {
-  MovieRequestPodo(num id) : super(id);
+class MovieRequest extends MediaContentRequest {
+  MovieRequest(num id) : super(id);
 
   Map<String, dynamic> toJson() => {
         'theMovieDbId': id,
