+++ lib/model/request/token_refresh.dart
@@ -1,7 +1,7 @@
-class TokenRefreshRequestPodo {
+class TokenRefreshRequest {
   String _token, _username;
 
-  TokenRefreshRequestPodo(this._token, this._username);
+  TokenRefreshRequest(this._token, this._username);
 
   Map<String, dynamic> toJson() => {
         'token': _token,
