+++ lib/pages/media_content/request_buttons.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:ombiapp/contracts/media_content.dart';
 import 'package:ombiapp/contracts/media_content_type.dart';
-import 'package:ombiapp/model/request/content/requests/movie.dart';
-import 'package:ombiapp/model/response/media_content/requests/media_content_request.dart';
+import 'package:ombiapp/model/request/content/requests/movie_request.dart';
+import 'package:ombiapp/model/response/media_content/requests/media_content_request_response.dart';
 import 'package:ombiapp/model/response/media_content/series/series.dart';
 import 'package:ombiapp/services/request_service.dart';
 import 'package:ombiapp/services/router.dart';
@@ -92,7 +92,7 @@ class _RequestButtonState extends State<RequestButton> {
           _searching = true;
         });
         requestManager.requestContent(
-            widget.content, MovieRequestPodo(widget.content.id));
+            widget.content, MovieRequest(widget.content.id));
         break;
       case MediaContentType.SERIES:
         RouterService.navigate(context, Routes.SERIES_REQUEST,
