+++ lib/services/network/content/request_content_bloc.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:ombiapp/contracts/media_content.dart';
 import 'package:ombiapp/contracts/media_content_request.dart';
-import 'package:ombiapp/model/response/media_content/requests/media_content_request.dart';
+import 'package:ombiapp/model/response/media_content/requests/media_content_request_response.dart';
 import 'package:ombiapp/services/network/repository.dart';
 import 'package:rxdart/rxdart.dart';
