+++ lib/services/network/repository.dart
@@ -2,10 +2,10 @@ import 'package:flutter/cupertino.dart';
 import 'package:ombiapp/contracts/media_content.dart';
 import 'package:ombiapp/contracts/media_content_request.dart';
 import 'package:ombiapp/contracts/media_content_type.dart';
-import 'package:ombiapp/model/request/login.dart';
-import 'package:ombiapp/model/response/LoginResponseDto.dart';
+import 'package:ombiapp/model/request/login_request.dart';
+import 'package:ombiapp/model/response/login_response.dart';
 import 'package:ombiapp/model/response/media_content/content_wrapper.dart';
-import 'package:ombiapp/model/response/media_content/requests/media_content_request.dart';
+import 'package:ombiapp/model/response/media_content/requests/media_content_request_response.dart';
 import 'package:ombiapp/model/response/user.dart';
 
 import 'api_provider.dart';
@@ -17,7 +17,7 @@ import 'api_provider.dart';
 abstract class RepositoryAPI {
   Future<User> getIdentity();
 
-  Future<LoginResponseDto> login(LoginRequestPodo loginRequestPodo);
+  Future<LoginResponseDto> login(LoginRequest loginRequestPodo);
 
   Future<bool> testConnection(String address);
   Future<ContentWrapper> contentSearch(
